refactor actual orchestrator impl out of orchestrator/mod.rs
